+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 7.4.1 (2019-04-02)
+
+### Bugfixes
+
+ - Fix custom input preservation during upgrade
+
 ## 7.4.0 (2019-04-01)
 
 ### Bugfixes